portfolio_evaluator/requirements.txt
fastapi>=0.100
uvicorn>=0.22
pydantic>=2.0
psycopg2-binary>=2.9
pandas>=2.0
scikit-learn>=1.2
joblib>=1.2
python-dotenv>=1.0
